+++ src/interpolator/one/mod.rs
@@ -291,3 +291,11 @@ where
         self.init_strategy()
     }
 }
+
+impl<T, S> DynInterpolator<T> for Interp1DOwned<T, S>
+where
+    T: Float + Euclid + Debug + Send + Sync + 'static,
+    S: Strategy1D<OwnedRepr<T>> + Clone + Send + Sync + 'static,
+{
+    dyn_interpolate_impl!();
+}