+++ src/interpolator/two/mod.rs
@@ -315,3 +315,11 @@ where
         self.init_strategy()
     }
 }
+
+impl<T, S> DynInterpolator<T> for Interp2DOwned<T, S>
+where
+    T: Float + Euclid + Debug + Send + Sync + 'static,
+    S: Strategy2D<OwnedRepr<T>> + Clone + Send + Sync + 'static,
+{
+    dyn_interpolate_impl!();
+}